Refactor and add test cases for new CommandExecutor's spiwith sh and py operator factories
